// server/data/interfaces/bedAssignment.ts
export interface BedAssignment {
  bookingId: number
  livingUnitId: number
  assignmentDate: string
  assignmentReason: string
  assignmentEndDate: string
  assignmentEndDateTime: string
  agencyId: string
  description: string
  bedAssignmentHistorySequence: number
  movementMadeBy: string
  offenderNo: string
}